v2

+++ Program.cs
@@ -19,7 +19,7 @@ namespace scvhost
             string s = "50m3th1n@Ph1$hy";
             foreach (char c in s)
             {
-                Console.WriteLine((int)c);
+                Console.WriteLine(c);
             };
